requirements.txt
Package          Version
---------------- -------
click            7.1.2
dnspython        1.16.0
eventlet         0.26.1
Flask            1.1.2
Flask-Login      0.5.0
Flask-SocketIO   4.3.1
Flask-SQLAlchemy 2.4.4
Flask-WTF        0.14.3
gevent           20.6.2
gevent-websocket 0.10.1
greenlet         0.4.16
gunicorn         20.0.4
itsdangerous     1.1.0
Jinja2           2.11.2
MarkupSafe       1.1.1
monotonic        1.5
passlib          1.7.2
pip              20.2
psycopg2-binary  2.8.5
python-dotenv    0.14.0
python-engineio  3.13.1
python-socketio  4.6.0
setuptools       41.6.0
six              1.15.0
SQLAlchemy       1.3.18
Werkzeug         1.0.1
WTForms          2.3.3
zope.event       4.4
zope.interface   5.1.0
